home page presentation and images moved
implemented the home page as much as it is done on figma now
