--- presentation/src/pages/home/home.js
@@ -1,20 +1,0 @@
-import './home.css';
-
-function Home() {
-    return (
-        <div className="home">
-            <img src="images/Biertan.jpg" alt='background'></img>
-
-            <div className="shortcuts">
-                <div className="shortcut-button">
-                    <div className="button-icon">
-
-                    </div>
-                    <p>Evenimente</p>
-                </div>
-            </div>
-        </div>
-    )
-}
-
-export default Home;
